improve configure script and update oh-my-zsh

+++ configure.sh
@@ -1,5 +1,5 @@
+#!/bin/bash
 # TODO: doc in readme
-
 git submodule update --init --recursive
 
 source ./install-deps.sh
@@ -17,18 +17,19 @@ echo "Dotfiles dir: $dotfiles_dir"
 
 # link zsh / oh-my-zsh configuration
 echo "Setting up shell..."
-ln -sf $dotfiles_dir/shell/oh-my-zsh $HOME/.oh-my-zsh
-ln -sf $dotfiles_dir/shell/zshrc $HOME/.zshrc
+
+[ ! -e $HOME/.oh-my-zsh ] && ln -sf $dotfiles_dir/shell/oh-my-zsh $HOME/.oh-my-zsh
+[ ! -e $HOME/.zshrc ] && ln -sf $dotfiles_dir/shell/zshrc $HOME/.zshrc
 
 # link vim config
 echo "Setting up vim config..."
-ln -sf $dotfiles_dir/vim $HOME/.vim
+[ ! -e $HOME/.vim ] && ln -sf $dotfiles_dir/vim $HOME/.vim
 vim -c "PlugInstall"
 
 # .git configuration
 echo "Setting up git config..."
-ln -sf $dotfiles_dir/git/gitignore_global $HOME/.gitignore_global
-ln -sf $dotfiles_dir/git/gitconfig.static $HOME/.gitconfig.static
+[ ! -e $HOME/.gitignore_global ] && ln -sf $dotfiles_dir/git/gitignore_global $HOME/.gitignore_global
+[ ! -e $HOME/.gitconfig.static ] && ln -sf $dotfiles_dir/git/gitconfig.static $HOME/.gitconfig.static
 git config --global include.path $HOME/.gitconfig.static
 echo 'Remember to setup user and email:'
 echo '    git config --global user.name "YOUR NAME"'
